refactor(db/seed): move brand data to model file
