--- libs/db/seed/Brands.ts
@@ -1,90 +1,0 @@
-export const brandData = [
-  {
-    name: 'cuHacking',
-    description: 'Carleton University\'s Official Hackathon',
-    mediaUrl:
-      'https://avatars.githubusercontent.com/u/29588588?s=400&u=8b806db4b4a6277abc155ed5d290dcfe6a6e7598&v=4',
-    links: [
-      { name: 'Website', link: 'https://cuhacking.ca' },
-      { name: 'Portal', link: 'https://portal.cuhacking.ca' },
-      { name: 'Design', link: 'https://design.cuhacking.ca/' },
-      { name: 'Architecture', link: 'https://arch.cuhacking.ca/view/index' },
-      {
-        name: 'Graph',
-        link: 'https://graph.cuhacking.ca/#/projects/all?groupByFolder=true',
-      },
-      {
-        name: 'Tooling',
-        link: 'https://docs.cuhacking.ca/contribution-guidelines/coding-standards/tooling',
-      },
-      { name: 'ESLint', link: 'https://eslint.cuhacking.ca/rules' },
-      { name: 'Discord', link: 'https://discord.gg/h2cQqF9aZf' },
-      { name: 'Instagram', link: 'https://www.instagram.com/cuhacking/' },
-      {
-        name: 'LinkedIn',
-        link: 'https://www.linkedin.com/company/cuhacking/',
-      },
-      { name: 'Linktree', link: 'https://linktr.ee/cuhacking_' },
-      {
-        name: 'Brand',
-        link: 'https://www.figma.com/design/wc1JOWR48tBNkjcjwY3AzB/%E2%8C%A8%EF%B8%8F-cuHacking-Design-System?node-id=0-1&t=YTR1ET4Qw1wG1cjz-1',
-      },
-      {
-        name: 'Project Board',
-        link: 'https://github.com/orgs/cuhacking/projects/4',
-      },
-      { name: 'GitHub', link: 'https://github.com/cuhacking/2025' },
-    ],
-  },
-  {
-    name: 'Carleton Computer Science Society',
-    description:
-      'Student-run society for Computer Science at Carleton University',
-    mediaUrl:
-      'https://carleton.ca/brand/wp-content/uploads/brand-logo-800w-1.jpg',
-    links: [{ name: 'Website', link: 'https://ccss.carleton.ca' }],
-  },
-  {
-    name: 'Software Engineering Student Association',
-    description:
-      'University of Ottawa\'s Software Engineering Student Association',
-    mediaUrl:
-      'https://www.uottawasesa.ca/static/media/logo.dc06f1167afbfbe0b7cf.png',
-    links: [{ name: 'Website', link: 'https://www.uottawasesa.ca' }],
-  },
-  {
-    name: 'Carleton University',
-    description: 'One of Canada\'s leading universities',
-    mediaUrl:
-      'https://cdn.carleton.ca/rds/assets/cu-logos/cu-logo-color-right-horiztonal.svg',
-    links: [{ name: 'Website', link: 'https://carleton.ca' }],
-  },
-  {
-    name: 'University of Ottawa',
-    description: 'Canada\'s university for bilingual education',
-    mediaUrl:
-      'https://images.seeklogo.com/logo-png/35/1/university-of-ottawa-logo-png_seeklogo-356159.png',
-    links: [{ name: 'Website', link: 'https://www.uottawa.ca/en' }],
-  },
-  {
-    name: 'BlackBerry QNX',
-    description: 'A leader in embedded systems and automotive software',
-    mediaUrl:
-      'https://blackberry.qnx.com/etc.clientlibs/bbcom/clientlibs/clientlib-etc-legacy/resources/bbcom-aem-project/images/BlackBerry-QNX-logo-white-new.png',
-    links: [{ name: 'Website', link: 'https://blackberry.qnx.com' }],
-  },
-  {
-    name: 'FullScript',
-    description:
-      'A platform for health professionals to support patient wellness',
-    mediaUrl:
-      'https://marketing-uploads.fullscript.com/wp-content/uploads/2024/01/24130952/fullscript-logo-r.svg',
-    links: [{ name: 'Website', link: 'https://fullscript.com' }],
-  },
-  {
-    name: 'Tail\'ed',
-    description: 'A marketplace for tailors and custom clothing',
-    mediaUrl: 'https://tailed.ca/tailed-logo-w-name.png',
-    links: [{ name: 'Website', link: 'https://tailed.ca' }],
-  },
-]
